assignment_3/pdf/requirements.txt
pymupdf
pdfplumber
pytesseract
Pillow
spacy
pandas
